requirements.txt
Pillow>=10.0.0
pillow-heif>=0.13.0
pymediainfo>=6.0.0
tqdm>=4.65.0
colorama>=0.4.6

